move formating utilitys to modular_pipeline_util.py
